requirements.txt
annotated-types>=0.7.0
anyio<=4.7.0
blinker<=1.9.0
click<=8.1.7
colorama<=0.4.6
Flask<=3.1.0
greenlet<=3.1.1
idna<=3.10
itsdangerous<=2.2.0
psycopg2-binary<=2.9.10
Jinja2<=3.1.4
MarkupSafe<=3.0.2
pip<=23.2.1
python-dotenv<=1.0.1
sniffio<=1.3.1
SQLAlchemy<=2.0.36
starlette<=0.41.3
typing_extensions<=4.12.2
Werkzeug<=3.1.3